Set up CI with bisys Azure Pipelines
Developing a full stack restaurant management system involves creating both the front-end and back-end components of the application. In this example, we'll create a simple system that includes features for managing menu items, orders, reservations, and customer information. We'll use Node.js for the back-end, Express.js for the server, MongoDB for the database, and React.js for the front-end. [skip ci]